requirements.txt
fastapi~=0.112.2
uvicorn
sqlalchemy~=1.4.39
psycopg2-binary
pydantic~=2.10.3
python-dotenv~=0.21.0
pytest~=7.4.0
httpx
azure-core~=1.33.0